+++ src/main/java/tcfplayz/chemicals/ChemicalsInit.java
@@ -13,11 +13,11 @@ import org.reflections.Reflections;
 
 import tcfplayz.chemicals.blocks.AtomCollider;
 import tcfplayz.chemicals.elements.Hydrogen;
-import tcfplayz.chemicals.items.Salt;
-import tcfplayz.chemicals.utils.blocks.Blocks;
-import tcfplayz.chemicals.utils.items.Elements;
-import tcfplayz.chemicals.utils.items.OtherItems;
-import tcfplayz.chemicals.utils.items.Solutions;
+import tcfplayz.chemicals.items.salt.Salt;
+import tcfplayz.misc.utils.blocks.Blocks;
+import tcfplayz.misc.utils.items.Elements;
+import tcfplayz.misc.utils.items.OtherItems;
+import tcfplayz.misc.utils.items.Solutions;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.Set;
