--- src/main/java/tcfplayz/chemicals/config/CreateConfig.java
@@ -1,4 +1,0 @@
-package tcfplayz.chemicals.config;
-
-public class CreateConfig {
-}
