+++ src/main/java/tcfplayz/chemicals/elements/Boron.java
@@ -2,14 +2,13 @@ package tcfplayz.chemicals.elements;
 
 import tcfplayz.misc.utils.items.Elements;
 
-public class Lithium extends Elements {
-    public Lithium() {
+public class Boron extends Elements {
+    public Boron() {
         super();
     }
 
     @Override
     public String getID() {
-        return "lithium";
+        return "boron";
     }
-
 }
