+++ src/main/java/tcfplayz/chemicals/elements/Carbon.java
@@ -2,14 +2,13 @@ package tcfplayz.chemicals.elements;
 
 import tcfplayz.misc.utils.items.Elements;
 
-public class Lithium extends Elements {
-    public Lithium() {
+public class Carbon extends Elements {
+    public Carbon() {
         super();
     }
 
     @Override
     public String getID() {
-        return "lithium";
+        return "carbon";
     }
-
 }
