+++ src/main/java/tcfplayz/chemicals/elements/Oxygen.java
@@ -2,14 +2,13 @@ package tcfplayz.chemicals.elements;
 
 import tcfplayz.misc.utils.items.Elements;
 
-public class Lithium extends Elements {
-    public Lithium() {
+public class Oxygen extends Elements {
+    public Oxygen() {
         super();
     }
 
     @Override
     public String getID() {
-        return "lithium";
+        return "oxygen";
     }
-
 }
