--- src/main/java/tcfplayz/chemicals/items/Salt.java
@@ -1,10 +1,0 @@
-package tcfplayz.chemicals.items;
-
-import tcfplayz.chemicals.utils.items.Solutions;
-
-public class Salt extends Solutions {
-    @Override
-    public String getID() {
-        return "salt";
-    }
-}
