+++ src/main/java/tcfplayz/chemicals/items/salt/Salt.java
@@ -1,0 +1,10 @@
+package tcfplayz.chemicals.items.salt;
+
+import tcfplayz.misc.utils.items.Solutions;
+
+public class Salt extends Solutions {
+    @Override
+    public String getID() {
+        return "salt";
+    }
+}
