--- src/main/java/tcfplayz/chemicals/tools/diminish/Diminisher.java
@@ -1,6 +1,0 @@
-package tcfplayz.chemicals.tools.diminish;
-
-public class Diminisher {
-    // hey players suffer more
-    // https://discord.com/channels/802077281165639691/914150012454318161/927159590330699796
-}
