--- src/main/java/tcfplayz/chemicals/tools/longwood/WoodenLongsword.java
@@ -1,4 +1,0 @@
-package tcfplayz.chemicals.tools.longwood;
-
-public class WoodenLongsword {
-}
