--- src/main/java/tcfplayz/chemicals/utils/items/Elements.java
@@ -1,15 +1,0 @@
-package tcfplayz.chemicals.utils.items;
-
-import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
-import net.minecraft.item.Item;
-
-import tcfplayz.chemicals.ChemicalsInit;
-
-public abstract class Elements extends Item {
-
-    public Elements() {
-        super(new FabricItemSettings().maxCount(64).group(ChemicalsInit.chemical));
-    }
-
-    public abstract String getID();
-}
