--- src/main/java/tcfplayz/chemicals/utils/items/OtherItems.java
@@ -1,13 +1,0 @@
-package tcfplayz.chemicals.utils.items;
-
-import net.minecraft.item.Item;
-import tcfplayz.chemicals.ChemicalsInit;
-
-public abstract class OtherItems extends Item {
-    public OtherItems() {
-        super(new Settings().group(ChemicalsInit.others).maxCount(16));
-    }
-
-    public abstract String getID();
-
-}
