--- src/main/java/tcfplayz/chemicals/utils/mobs/BossCreator.java
@@ -1,13 +1,0 @@
-package tcfplayz.chemicals.utils.mobs;
-
-import net.minecraft.entity.EntityType;
-import net.minecraft.entity.mob.PathAwareEntity;
-import net.minecraft.world.World;
-
-public abstract class BossCreator extends PathAwareEntity {
-    protected BossCreator(EntityType<? extends PathAwareEntity> entityType, World world) {
-        super(entityType, world);
-    }
-
-    abstract public String getID();
-}
