--- src/main/java/tcfplayz/chemicals/utils/other/Timeout.java
@@ -1,15 +1,0 @@
-package tcfplayz.chemicals.utils.other;
-
-public class Timeout {
-    public static void setTimeout(Runnable runnable, int delay){
-        new Thread(() -> {
-            try {
-                Thread.sleep(delay);
-                runnable.run();
-            }
-            catch (Exception e){
-                System.err.println(e);
-            }
-        }).start();
-    }
-}
