--- src/main/java/tcfplayz/chemicals/vanillaplus/LightningRod.java
@@ -1,9 +1,0 @@
-package tcfplayz.chemicals.vanillaplus;
-
-import net.minecraft.block.LightningRodBlock;
-
-public class LightningRod extends LightningRodBlock {
-    public LightningRod(Settings settings) {
-        super(settings);
-    }
-}
