+++ src/main/java/tcfplayz/misc/biome/beedim/BeeDimChunkGen.java
@@ -1,0 +1,132 @@
+package tcfplayz.misc.biome.beedim;
+
+/*
+import com.mojang.serialization.Codec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.minecraft.block.BlockState;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.ChunkRegion;
+import net.minecraft.world.HeightLimitView;
+import net.minecraft.world.Heightmap;
+import net.minecraft.world.biome.source.BiomeAccess;
+import net.minecraft.world.biome.source.BiomeSource;
+import net.minecraft.world.biome.source.util.MultiNoiseUtil;
+import net.minecraft.world.chunk.Chunk;
+import net.minecraft.world.gen.GenerationStep;
+import net.minecraft.world.gen.StructureAccessor;
+import net.minecraft.world.gen.chunk.Blender;
+import net.minecraft.world.gen.chunk.ChunkGenerator;
+import net.minecraft.world.gen.chunk.VerticalBlockSample;
+import net.minecraft.world.gen.noise.NoiseConfig;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.Executor;
+*/
+
+public class BeeDimChunkGen { /* TODO: dimension
+extends ChunkGenerator {
+
+    public static final Codec<BeeDimChunkGen> CODEC = RecordCodecBuilder.create((instance) ->
+            instance.group(
+                    BiomeSource.CODEC.fieldOf("biome_source")
+                            .forGetter((generator) -> generator.biomeSource)
+            ).apply(instance, instance.stable(BeeDimChunkGen::new))
+    );
+
+    public BeeDimChunkGen(BiomeSource biomeSource) {
+        super(biomeSource, new StructuresConfig(Optional.empty(), Collections.emptyMap()));
+    }
+
+    @Override
+    protected Codec<? extends ChunkGenerator> getCodec() {
+        return CODEC;
+    }
+
+    @Override
+    public void carve(ChunkRegion chunkRegion, long seed, NoiseConfig noiseConfig, BiomeAccess world, StructureAccessor structureAccessor, Chunk chunk, GenerationStep.Carver carverStep) {
+
+    }
+
+    @Override
+    public void buildSurface(ChunkRegion region, StructureAccessor structures, NoiseConfig noiseConfig, Chunk chunk) {
+
+    }
+
+    @Override
+    public ChunkGenerator withSeed(long seed) {
+        return this;
+    }
+
+    @Override
+    public MultiNoiseUtil.MultiNoiseSampler getMultiNoiseSampler() {
+        return null;
+    }
+
+    @Override
+    public void carve(ChunkRegion chunkRegion, long seed, BiomeAccess biomeAccess, StructureAccessor structureAccessor, Chunk chunk, GenerationStep.Carver generationStep) {
+
+    }
+
+    @Override
+    public void buildSurface(ChunkRegion region, StructureAccessor structures, Chunk chunk) {
+
+    }
+
+    @Override
+    public void populateEntities(ChunkRegion region) {
+
+    }
+
+    @Override
+    public int getWorldHeight() {
+        return 1024;
+    }
+
+    @Override
+    public CompletableFuture<Chunk> populateNoise(Executor executor, Blender blender, NoiseConfig noiseConfig, StructureAccessor structureAccessor, Chunk chunk) {
+        return null;
+    }
+
+    @Override
+    public CompletableFuture<Chunk> populateNoise(Executor executor, Blender blender, StructureAccessor structureAccessor, Chunk chunk) {
+        return null;
+    }
+
+    @Override
+    public int getSeaLevel() {
+        return 120;
+    }
+
+    @Override
+    public int getMinimumY() {
+        return -320;
+    }
+
+    @Override
+    public int getHeight(int x, int z, Heightmap.Type heightmap, HeightLimitView world, NoiseConfig noiseConfig) {
+        return 0;
+    }
+
+    @Override
+    public VerticalBlockSample getColumnSample(int x, int z, HeightLimitView world, NoiseConfig noiseConfig) {
+        return null;
+    }
+
+    @Override
+    public void getDebugHudText(List<String> text, NoiseConfig noiseConfig, BlockPos pos) {
+
+    }
+
+    @Override
+    public int getHeight(int x, int z, Heightmap.Type heightmap, HeightLimitView world) {
+        return 1024;
+    }
+
+    @Override
+    public VerticalBlockSample getColumnSample(int x, int z, HeightLimitView world) {
+        return new VerticalBlockSample(-320, new BlockState[0]);
+    } */
+}
