+++ src/main/java/tcfplayz/misc/tools/longwood/WoodenLongsword.java
@@ -1,0 +1,4 @@
+package tcfplayz.misc.tools.longwood;
+
+public class WoodenLongsword {
+}
