+++ src/main/java/tcfplayz/misc/tools/phobictools/PhobicTool.java
@@ -1,0 +1,11 @@
+package tcfplayz.misc.tools.phobictools;
+
+import tcfplayz.chemicals.elements.Hydrogen;
+import tcfplayz.misc.utils.tools.Tools;
+
+public class PhobicTool extends Tools {
+
+    public PhobicTool() {
+        super(new PhobicTool(), 5000, 15.0f, 15.0f, 75, 100, new Hydrogen());
+    }
+}
