+++ src/main/java/tcfplayz/misc/utils/blocks/Blocks.java
@@ -1,0 +1,13 @@
+package tcfplayz.misc.utils.blocks;
+
+import net.minecraft.block.Block;
+import net.minecraft.block.Material;
+
+public abstract class Blocks extends Block {
+    public Blocks() {
+        super(Block.Settings.of(Material.METAL).requiresTool().hardness(4f));
+    }
+
+    public abstract String getID();
+
+}
