+++ src/main/java/tcfplayz/misc/utils/items/Elements.java
@@ -1,0 +1,15 @@
+package tcfplayz.misc.utils.items;
+
+import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
+import net.minecraft.item.Item;
+
+import tcfplayz.chemicals.ChemicalsInit;
+
+public abstract class Elements extends Item {
+
+    public Elements() {
+        super(new FabricItemSettings().maxCount(64).group(ChemicalsInit.chemical));
+    }
+
+    public abstract String getID();
+}
