+++ src/main/java/tcfplayz/misc/utils/items/OtherItems.java
@@ -1,0 +1,13 @@
+package tcfplayz.misc.utils.items;
+
+import net.minecraft.item.Item;
+import tcfplayz.chemicals.ChemicalsInit;
+
+public abstract class OtherItems extends Item {
+    public OtherItems() {
+        super(new Settings().group(ChemicalsInit.others).maxCount(16));
+    }
+
+    public abstract String getID();
+
+}
