+++ src/main/java/tcfplayz/misc/utils/items/Solutions.java
@@ -1,0 +1,13 @@
+package tcfplayz.misc.utils.items;
+
+import net.minecraft.item.Item;
+import tcfplayz.chemicals.ChemicalsInit;
+
+public abstract class Solutions extends Item {
+    public Solutions() {
+        super(new Item.Settings().maxCount(16).group(ChemicalsInit.chemical));
+    }
+
+    abstract public String getID();
+
+}
