+++ src/main/java/tcfplayz/misc/utils/mobs/BossCreator.java
@@ -1,0 +1,13 @@
+package tcfplayz.misc.utils.mobs;
+
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.mob.PathAwareEntity;
+import net.minecraft.world.World;
+
+public abstract class BossCreator extends PathAwareEntity {
+    protected BossCreator(EntityType<? extends PathAwareEntity> entityType, World world) {
+        super(entityType, world);
+    }
+
+    abstract public String getID();
+}
