+++ src/main/java/tcfplayz/misc/utils/other/Timeout.java
@@ -1,0 +1,15 @@
+package tcfplayz.misc.utils.other;
+
+public class Timeout {
+    public static void setTimeout(Runnable runnable, int delay){
+        new Thread(() -> {
+            try {
+                Thread.sleep(delay);
+                runnable.run();
+            }
+            catch (Exception e){
+                System.err.println(e);
+            }
+        }).start();
+    }
+}
